refactor: sprite renderable and text renderable
