+++ packages/core/src/2d/sprite/SpriteRenderer.ts
@@ -1,149 +1,29 @@
-import { BoundingBox, Color, MathUtil } from "@galacean/engine-math";
+import { Color } from "@galacean/engine-math";
 import { Entity } from "../../Entity";
-import { BatchUtils } from "../../RenderPipeline/BatchUtils";
 import { PrimitiveChunkManager } from "../../RenderPipeline/PrimitiveChunkManager";
 import { RenderContext } from "../../RenderPipeline/RenderContext";
 import { SubPrimitiveChunk } from "../../RenderPipeline/SubPrimitiveChunk";
-import { SubRenderElement } from "../../RenderPipeline/SubRenderElement";
 import { Renderer, RendererUpdateFlags } from "../../Renderer";
-import { assignmentClone, deepClone, ignoreClone } from "../../clone/CloneManager";
+import { deepClone, ignoreClone } from "../../clone/CloneManager";
+import { SpriteMaskLayer } from "../../enums/SpriteMaskLayer";
+import { Material } from "../../material";
 import { ShaderProperty } from "../../shader/ShaderProperty";
-import { ISpriteAssembler } from "../assembler/ISpriteAssembler";
-import { ISpriteRenderer } from "../assembler/ISpriteRenderer";
-import { SimpleSpriteAssembler } from "../assembler/SimpleSpriteAssembler";
-import { SlicedSpriteAssembler } from "../assembler/SlicedSpriteAssembler";
-import { TiledSpriteAssembler } from "../assembler/TiledSpriteAssembler";
+import { Texture2D } from "../../texture";
+import { ISpriteLayout } from "./ISpriteLayout";
 import { SpriteDrawMode } from "../enums/SpriteDrawMode";
 import { SpriteMaskInteraction } from "../enums/SpriteMaskInteraction";
-import { SpriteModifyFlags } from "../enums/SpriteModifyFlags";
-import { SpriteTileMode } from "../enums/SpriteTileMode";
-import { Sprite } from "./Sprite";
-import { SpriteMaskLayer } from "../../enums/SpriteMaskLayer";
+import { SpriteRenderable, SpriteRenderableFlags } from "./SpriteRenderable";
+import { WorldSpriteLayout } from "./WorldSpriteLayout";
 
 /**
  * Renders a Sprite for 2D graphics.
  */
-export class SpriteRenderer extends Renderer implements ISpriteRenderer {
+export class SpriteRenderer extends SpriteRenderable(Renderer) {
   /** @internal */
   static _textureProperty: ShaderProperty = ShaderProperty.getByName("renderer_SpriteTexture");
 
-  /** @internal */
-  @ignoreClone
-  _subChunk: SubPrimitiveChunk;
-
-  @ignoreClone
-  private _drawMode: SpriteDrawMode;
-  @ignoreClone
-  private _assembler: ISpriteAssembler;
-  @assignmentClone
-  private _tileMode: SpriteTileMode = SpriteTileMode.Continuous;
-  @assignmentClone
-  private _tiledAdaptiveThreshold: number = 0.5;
-
   @deepClone
   private _color: Color = new Color(1, 1, 1, 1);
-  @ignoreClone
-  private _sprite: Sprite = null;
-
-  @ignoreClone
-  private _automaticWidth: number = 0;
-  @ignoreClone
-  private _automaticHeight: number = 0;
-  @assignmentClone
-  private _customWidth: number = undefined;
-  @assignmentClone
-  private _customHeight: number = undefined;
-  @assignmentClone
-  private _flipX: boolean = false;
-  @assignmentClone
-  private _flipY: boolean = false;
-
-  /**
-   * The draw mode of the sprite renderer.
-   */
-  get drawMode(): SpriteDrawMode {
-    return this._drawMode;
-  }
-
-  set drawMode(value: SpriteDrawMode) {
-    if (this._drawMode !== value) {
-      this._drawMode = value;
-      switch (value) {
-        case SpriteDrawMode.Simple:
-          this._assembler = SimpleSpriteAssembler;
-          break;
-        case SpriteDrawMode.Sliced:
-          this._assembler = SlicedSpriteAssembler;
-          break;
-        case SpriteDrawMode.Tiled:
-          this._assembler = TiledSpriteAssembler;
-          break;
-        default:
-          break;
-      }
-      this._assembler.resetData(this);
-      this._dirtyUpdateFlag |= SpriteRendererUpdateFlags.WorldVolumeUVAndColor;
-    }
-  }
-
-  /**
-   * The tiling mode of the sprite renderer. (Only works in tiled mode.)
-   */
-  get tileMode(): SpriteTileMode {
-    return this._tileMode;
-  }
-
-  set tileMode(value: SpriteTileMode) {
-    if (this._tileMode !== value) {
-      this._tileMode = value;
-      if (this.drawMode === SpriteDrawMode.Tiled) {
-        this._dirtyUpdateFlag |= SpriteRendererUpdateFlags.WorldVolumeUVAndColor;
-      }
-    }
-  }
-
-  /**
-   * Stretch Threshold in Tile Adaptive Mode, specified in normalized. (Only works in tiled adaptive mode.)
-   */
-  get tiledAdaptiveThreshold(): number {
-    return this._tiledAdaptiveThreshold;
-  }
-
-  set tiledAdaptiveThreshold(value: number) {
-    if (value !== this._tiledAdaptiveThreshold) {
-      value = MathUtil.clamp(value, 0, 1);
-      this._tiledAdaptiveThreshold = value;
-      if (this.drawMode === SpriteDrawMode.Tiled) {
-        this._dirtyUpdateFlag |= SpriteRendererUpdateFlags.WorldVolumeUVAndColor;
-      }
-    }
-  }
-
-  /**
-   * The Sprite to render.
-   */
-  get sprite(): Sprite {
-    return this._sprite;
-  }
-
-  set sprite(value: Sprite | null) {
-    const lastSprite = this._sprite;
-    if (lastSprite !== value) {
-      if (lastSprite) {
-        this._addResourceReferCount(lastSprite, -1);
-        lastSprite._updateFlagManager.removeListener(this._onSpriteChange);
-      }
-      this._dirtyUpdateFlag |= SpriteRendererUpdateFlags.All;
-      if (value) {
-        this._addResourceReferCount(value, 1);
-        value._updateFlagManager.addListener(this._onSpriteChange);
-        this.shaderData.setTexture(SpriteRenderer._textureProperty, value.texture);
-      } else {
-        this.shaderData.setTexture(SpriteRenderer._textureProperty, null);
-      }
-      this._sprite = value;
-    }
-  }
 
   /**
    * Rendering color for the Sprite graphic.
@@ -166,20 +46,16 @@ export class SpriteRenderer extends Renderer implements ISpriteRenderer {
    * otherwise return `SpriteRenderer.sprite.width`.
    */
   get width(): number {
-    if (this._customWidth !== undefined) {
-      return this._customWidth;
-    } else {
-      this._dirtyUpdateFlag & SpriteRendererUpdateFlags.AutomaticSize && this._calDefaultSize();
-      return this._automaticWidth;
-    }
+    return (<WorldSpriteLayout>this._layout).width;
   }
 
   set width(value: number) {
-    if (this._customWidth !== value) {
-      this._customWidth = value;
+    const layout = <WorldSpriteLayout>this._layout;
+    if (layout.customWidth !== value) {
+      layout.width = value;
       this._dirtyUpdateFlag |=
-        this._drawMode === SpriteDrawMode.Tiled
-          ? SpriteRendererUpdateFlags.WorldVolumeUVAndColor
+        this.drawMode === SpriteDrawMode.Tiled
+          ? SpriteRenderableFlags.WorldVolumeUVAndColor
           : RendererUpdateFlags.WorldVolume;
     }
   }
@@ -192,20 +68,16 @@ export class SpriteRenderer extends Renderer implements ISpriteRenderer {
    * otherwise return `SpriteRenderer.sprite.height`.
    */
   get height(): number {
-    if (this._customHeight !== undefined) {
-      return this._customHeight;
-    } else {
-      this._dirtyUpdateFlag & SpriteRendererUpdateFlags.AutomaticSize && this._calDefaultSize();
-      return this._automaticHeight;
-    }
+    return (<WorldSpriteLayout>this._layout).height;
   }
 
   set height(value: number) {
-    if (this._customHeight !== value) {
-      this._customHeight = value;
+    const layout = <WorldSpriteLayout>this._layout;
+    if (layout.customHeight !== value) {
+      layout.height = value;
       this._dirtyUpdateFlag |=
-        this._drawMode === SpriteDrawMode.Tiled
-          ? SpriteRendererUpdateFlags.WorldVolumeUVAndColor
+        this.drawMode === SpriteDrawMode.Tiled
+          ? SpriteRenderableFlags.WorldVolumeUVAndColor
           : RendererUpdateFlags.WorldVolume;
     }
   }
@@ -214,12 +86,13 @@ export class SpriteRenderer extends Renderer implements ISpriteRenderer {
    * Flips the sprite on the X axis.
    */
   get flipX(): boolean {
-    return this._flipX;
+    return (<WorldSpriteLayout>this._layout).flipX;
   }
 
   set flipX(value: boolean) {
-    if (this._flipX !== value) {
-      this._flipX = value;
+    const layout = <WorldSpriteLayout>this._layout;
+    if (layout.flipX !== value) {
+      layout.flipX = value;
       this._dirtyUpdateFlag |= RendererUpdateFlags.WorldVolume;
     }
   }
@@ -228,12 +101,13 @@ export class SpriteRenderer extends Renderer implements ISpriteRenderer {
    * Flips the sprite on the Y axis.
    */
   get flipY(): boolean {
-    return this._flipY;
+    return (<WorldSpriteLayout>this._layout).flipY;
   }
 
   set flipY(value: boolean) {
-    if (this._flipY !== value) {
-      this._flipY = value;
+    const layout = <WorldSpriteLayout>this._layout;
+    if (layout.flipY !== value) {
+      layout.flipY = value;
       this._dirtyUpdateFlag |= RendererUpdateFlags.WorldVolume;
     }
   }
@@ -267,228 +141,49 @@ export class SpriteRenderer extends Renderer implements ISpriteRenderer {
    */
   constructor(entity: Entity) {
     super(entity);
-    this.drawMode = SpriteDrawMode.Simple;
-    this._dirtyUpdateFlag |= SpriteRendererUpdateFlags.Color;
-    this.setMaterial(this._engine._basicResources.spriteDefaultMaterial);
-    this._onSpriteChange = this._onSpriteChange.bind(this);
+    this._initSpriteRenderable(SpriteRenderer._textureProperty);
     //@ts-ignore
     this._color._onValueChanged = this._onColorChanged.bind(this);
   }
 
-  /**
-   * @internal
-   */
-  override _updateTransformShaderData(context: RenderContext, onlyMVP: boolean, batched: boolean): void {
-    //@todo: Always update world positions to buffer, should opt
-    super._updateTransformShaderData(context, onlyMVP, true);
-  }
+  // ===== Abstract implementations =====
 
-  /**
-   * @internal
-   */
-  override _cloneTo(target: SpriteRenderer): void {
-    super._cloneTo(target);
-    target.sprite = this._sprite;
-    target.drawMode = this._drawMode;
-  }
-
-  /**
-   * @internal
-   */
-  override _canBatch(elementA: SubRenderElement, elementB: SubRenderElement): boolean {
-    return BatchUtils.canBatchSprite(elementA, elementB);
-  }
-
-  /**
-   * @internal
-   */
-  override _batch(elementA: SubRenderElement, elementB?: SubRenderElement): void {
-    BatchUtils.batchFor2D(elementA, elementB);
-  }
-
-  /**
-   * @internal
-   */
-  _getChunkManager(): PrimitiveChunkManager {
+  /** @internal */
+  override _getChunkManager(): PrimitiveChunkManager {
     return this.engine._batcherManager.primitiveChunkManager2D;
   }
 
-  protected override _updateBounds(worldBounds: BoundingBox): void {
-    const sprite = this._sprite;
-    if (sprite) {
-      this._assembler.updatePositions(
-        this,
-        this._transformEntity.transform.worldMatrix,
-        this.width,
-        this.height,
-        sprite.pivot,
-        this._flipX,
-        this._flipY
-      );
-    } else {
-      const { worldPosition } = this._transformEntity.transform;
-      worldBounds.min.copyFrom(worldPosition);
-      worldBounds.max.copyFrom(worldPosition);
-    }
+  /** @internal */
+  override _getDefaultSpriteMaterial(): Material {
+    return this._engine._basicResources.spriteDefaultMaterial;
   }
 
-  protected override _render(context: RenderContext): void {
-    const { _sprite: sprite } = this;
-    if (!sprite?.texture || !this.width || !this.height) {
-      return;
-    }
-
-    let material = this.getMaterial();
-    if (!material) {
-      return;
-    }
-    // @todo: This question needs to be raised rather than hidden.
-    if (material.destroyed) {
-      material = this._engine._basicResources.spriteDefaultMaterial;
-    }
-
-    // Update position
-    if (this._dirtyUpdateFlag & RendererUpdateFlags.WorldVolume) {
-      this._assembler.updatePositions(
-        this,
-        this._transformEntity.transform.worldMatrix,
-        this.width,
-        this.height,
-        sprite.pivot,
-        this._flipX,
-        this._flipY
-      );
-      this._dirtyUpdateFlag &= ~RendererUpdateFlags.WorldVolume;
-    }
-
-    // Update uv
-    if (this._dirtyUpdateFlag & SpriteRendererUpdateFlags.UV) {
-      this._assembler.updateUVs(this);
-      this._dirtyUpdateFlag &= ~SpriteRendererUpdateFlags.UV;
-    }
-
-    // Update color
-    if (this._dirtyUpdateFlag & SpriteRendererUpdateFlags.Color) {
-      this._assembler.updateColor(this, 1);
-      this._dirtyUpdateFlag &= ~SpriteRendererUpdateFlags.Color;
-    }
-
-    // Push primitive
+  /** @internal */
+  override _submitSpriteRenderElement(
+    context: RenderContext,
+    material: Material,
+    subChunk: SubPrimitiveChunk,
+    texture: Texture2D
+  ): void {
     const camera = context.camera;
     const engine = camera.engine;
     const renderElement = engine._renderElementPool.get();
     renderElement.set(this.priority, this._distanceForSort);
     const subRenderElement = engine._subRenderElementPool.get();
-    const subChunk = this._subChunk;
-    subRenderElement.set(this, material, subChunk.chunk.primitive, subChunk.subMesh, this.sprite.texture, subChunk);
+    subRenderElement.set(this, material, subChunk.chunk.primitive, subChunk.subMesh, texture, subChunk);
     renderElement.addSubRenderElement(subRenderElement);
     camera._renderPipeline.pushRenderElement(context, renderElement);
   }
 
-  protected override _onDestroy(): void {
-    const sprite = this._sprite;
-    if (sprite) {
-      this._addResourceReferCount(sprite, -1);
-      sprite._updateFlagManager.removeListener(this._onSpriteChange);
-    }
-
-    super._onDestroy();
-
-    this._sprite = null;
-    this._assembler = null;
-    if (this._subChunk) {
-      this._getChunkManager().freeSubChunk(this._subChunk);
-      this._subChunk = null;
-    }
+  /** @internal */
+  override _createLayout(): ISpriteLayout {
+    return new WorldSpriteLayout(() => this.sprite);
   }
 
-  private _calDefaultSize(): void {
-    const sprite = this._sprite;
-    if (sprite) {
-      this._automaticWidth = sprite.width;
-      this._automaticHeight = sprite.height;
-    } else {
-      this._automaticWidth = this._automaticHeight = 0;
-    }
-    this._dirtyUpdateFlag &= ~SpriteRendererUpdateFlags.AutomaticSize;
-  }
-
-  @ignoreClone
-  private _onSpriteChange(type: SpriteModifyFlags): void {
-    switch (type) {
-      case SpriteModifyFlags.texture:
-        this.shaderData.setTexture(SpriteRenderer._textureProperty, this.sprite.texture);
-        break;
-      case SpriteModifyFlags.size:
-        this._dirtyUpdateFlag |= SpriteRendererUpdateFlags.AutomaticSize;
-        if (this._customWidth === undefined || this._customHeight === undefined) {
-          this._dirtyUpdateFlag |= RendererUpdateFlags.WorldVolume;
-        }
-        switch (this._drawMode) {
-          case SpriteDrawMode.Simple:
-            // When the width and height of `SpriteRenderer` are `undefined`,
-            // the `size` of `Sprite` will affect the position of `SpriteRenderer`.
-            if (this._customWidth === undefined || this._customHeight === undefined) {
-              this._dirtyUpdateFlag |= RendererUpdateFlags.WorldVolume;
-            }
-            break;
-          case SpriteDrawMode.Sliced:
-            this._dirtyUpdateFlag |= RendererUpdateFlags.WorldVolume;
-            break;
-          case SpriteDrawMode.Tiled:
-            this._dirtyUpdateFlag |= SpriteRendererUpdateFlags.WorldVolumeUVAndColor;
-            break;
-        }
-        break;
-      case SpriteModifyFlags.border:
-        switch (this._drawMode) {
-          case SpriteDrawMode.Sliced:
-            this._dirtyUpdateFlag |= SpriteRendererUpdateFlags.WorldVolumeAndUV;
-            break;
-          case SpriteDrawMode.Tiled:
-            this._dirtyUpdateFlag |= SpriteRendererUpdateFlags.WorldVolumeUVAndColor;
-            break;
-          default:
-            break;
-        }
-        break;
-      case SpriteModifyFlags.region:
-      case SpriteModifyFlags.atlasRegionOffset:
-        this._dirtyUpdateFlag |= SpriteRendererUpdateFlags.WorldVolumeAndUV;
-        break;
-      case SpriteModifyFlags.atlasRegion:
-        this._dirtyUpdateFlag |= SpriteRendererUpdateFlags.UV;
-        break;
-      case SpriteModifyFlags.pivot:
-        this._dirtyUpdateFlag |= RendererUpdateFlags.WorldVolume;
-        break;
-      case SpriteModifyFlags.destroy:
-        this.sprite = null;
-        break;
-    }
-  }
+  // ===== Private =====
 
   @ignoreClone
   private _onColorChanged(): void {
-    this._dirtyUpdateFlag |= SpriteRendererUpdateFlags.Color;
+    this._dirtyUpdateFlag |= SpriteRenderableFlags.Color;
   }
-}
-
-/**
- * @remarks Extends `RendererUpdateFlags`.
- */
-enum SpriteRendererUpdateFlags {
-  /** UV. */
-  UV = 0x2,
-  /** Color. */
-  Color = 0x4,
-  /** Automatic Size. */
-  AutomaticSize = 0x8,
-
-  /** WorldVolume and UV. */
-  WorldVolumeAndUV = 0x3,
-  /** WorldVolume, UV and Color. */
-  WorldVolumeUVAndColor = 0x7,
-  /** All. */
-  All = 0xf
 }
