+++ packages/core/src/2d/sprite/index.ts
@@ -1,3 +1,9 @@
+export type { ISpriteLayout } from "./ISpriteLayout";
 export { Sprite } from "./Sprite";
+export { SpriteDataBinding } from "./SpriteDataBinding";
+export type { ISpriteDataBindingOwner } from "./SpriteDataBinding";
+export type { ISpriteRenderable } from "./SpriteRenderable";
+export { SpriteRenderable, SpriteRenderableFlags } from "./SpriteRenderable";
 export { SpriteMask } from "./SpriteMask";
 export { SpriteRenderer } from "./SpriteRenderer";
+export { WorldSpriteLayout } from "./WorldSpriteLayout";
