+++ packages/core/src/2d/text/index.ts
@@ -1,5 +1,8 @@
 export { Font } from "./Font";
+export type { ITextRenderable } from "./TextRenderable";
+export { TextChunk, TextRenderable, TextRenderableFlags } from "./TextRenderable";
 export { TextRenderer } from "./TextRenderer";
+export { WorldTextLayout } from "./WorldTextLayout";
 // For set TextUtils._extendHeight used to extend the height of canvas, because in miniprogram performance is different from h5
 export { CharRenderInfo } from "./CharRenderInfo";
 export { SubFont } from "./SubFont";
