+++ packages/core/src/RenderPipeline/index.ts
@@ -1,5 +1,8 @@
 export { BasicRenderPipeline, RenderQueueFlags } from "./BasicRenderPipeline";
 export { BatchUtils } from "./BatchUtils";
 export { Blitter } from "./Blitter";
+export { PrimitiveChunkManager } from "./PrimitiveChunkManager";
+export { RenderContext } from "./RenderContext";
 export { RenderQueue } from "./RenderQueue";
+export { SubPrimitiveChunk } from "./SubPrimitiveChunk";
 export { PipelineStage } from "./enums/PipelineStage";
